pyproject.toml
[project]
name = "openarg"
version = "0.1.0"
description = "Inteligencia sobre Datos Abiertos de Argentina - Multi-Agent Backend"
requires-python = ">=3.12"
license = "MIT"
readme = "README.md"

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "uvloop>=0.21.0",
    "orjson>=3.10.0",
    "websockets>=14.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.14.0",
    "pgvector>=0.2.5",

    # DI
    "dishka>=1.6.0",

    # Workers
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",

    # AI / LLM
    "google-generativeai>=0.8.0",
    "anthropic>=0.40.0",

    # Data Processing
    "python-dateutil>=2.9.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
    "odfpy>=1.4.1",
    "fiona>=1.10.0",
    "httpx>=0.28.0",
    "aiofiles>=24.1.0",

    # YAML config (used by marts).
    "pyyaml>=6.0",

    # AWS
    "boto3>=1.28.0",

    # Config
    "pydantic[email]>=2.10.0",
    "tomli>=2.0.0;python_version<'3.11'",

    # Security
    "bcrypt>=4.3.0",
    "PyJWT[crypto]>=2.10.0",

    # Rate limiting
    "slowapi>=0.1.9",

    # SQL validation
    "sqlglot>=26.0.0",

    # Monitoring
    "prometheus-client>=0.21.0",

    # Structured logging & error tracking
    "structlog>=24.0",
    "sentry-sdk[fastapi]>=2.0",

    # Agent orchestration
    "langgraph>=1.0.0",
    "langgraph-checkpoint-postgres>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-timeout>=2.3.0",
    "pytest-rerunfailures>=14.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
]

# MASTERPLAN Fase 5 — dbt is OPTIONAL. The medallion runs end-to-end
# without dbt via the Celery tasks in `application.marts`. Install this
# extra (`pip install openarg[dbt]`) only when you want lineage / tests /
# docs from dbt on top.
dbt = [
    "dbt-core>=1.8.0",
    "dbt-postgres>=1.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]
# Disabled rules:
# - TC: Dishka DI + FastAPI require runtime imports
# - B: too noisy (B904 raise-from on 48 files)
# - N: naming conventions conflict with existing code
# - SIM: style opinions, not bugs

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
check_untyped_defs = true
warn_unused_configs = true
warn_unused_ignores = false
warn_return_any = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "e2e: end-to-end tests requiring real infrastructure (staging DB, LLMs)",
]
addopts = "-m 'not e2e'"
